+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+October 2024
+
+- Added indicator when dragging using copy mode (ctrl+drag or ctrl+alt+drag).
+
 September 2024
 
 - Added side, group and unit locking to prevent accidental changes.